api/requirements.txt
Django==3.2.9
psycopg2-binary==2.9.1
gunicorn==20.1.0
djangorestframework==3.12.4
whitenoise==5.3.0
django-cors-headers==3.10.0